+++ lib/bootsnap.rb
@@ -13,7 +13,7 @@ module Bootsnap
     development_mode: true,
     load_path_cache: true,
     autoload_paths_cache: nil,
-    disable_trace: false,
+    disable_trace: nil,
     compile_cache_iseq: true,
     compile_cache_yaml: true
   )
@@ -22,7 +22,11 @@ module Bootsnap
         "If you use Zeitwerk this option is useless, and if you are still using the classic autoloader " \
         "upgrading is recommended."
     end
-    setup_disable_trace if disable_trace
+
+    unless disable_trace.nil?
+      warn "[DEPRECATED] Bootsnap's `disable_trace:` option is deprecated and will be removed. " \
+        "If you use Ruby 2.5 or newer this option is useless, if not upgrading is recommended."
+    end
 
     Bootsnap::LoadPathCache.setup(
       cache_path:       cache_dir + '/bootsnap/load-path-cache',
@@ -34,16 +38,5 @@ module Bootsnap
       iseq: compile_cache_iseq,
       yaml: compile_cache_yaml
     )
-  end
-
-  def self.setup_disable_trace
-    if Gem::Version.new(RUBY_VERSION) >= Gem::Version.new('2.5.0')
-      warn(
-        "from #{caller_locations(1, 1)[0]}: The 'disable_trace' method is not allowed with this Ruby version. " \
-        "current: #{RUBY_VERSION}, allowed version: < 2.5.0",
-      )
-    else
-      RubyVM::InstructionSequence.compile_option = { trace_instruction: false }
-    end
   end
 end