+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Fix detection of YAML files in gems.
 * Adds an instrumentation API to monitor cache misses.
 * Allow to control the behavior of `require 'bootsnap/setup'` using environment variables.
 * Deprecate the `disable_trace` option.